Migrate book_creates_reasonable_search_index to BookTest
