+++ tests/testsuite/search/reasonable_search_index/src/SUMMARY.md
@@ -1,0 +1,10 @@
+# Summary
+
+[Introduction](intro.md)
+
+- [First Chapter](first/index.md)
+    - [Includes](first/includes.md)
+    - [Unicode](first/unicode.md)
+    - [No Headers](first/no-headers.md)
+    - [Duplicate Headers](first/duplicate-headers.md)
+    - [Heading Attributes](first/heading-attributes.md)
