+++ tests/testsuite/search/reasonable_search_index/src/first/duplicate-headers.md
@@ -1,0 +1,9 @@
+# Duplicate headers
+
+This page validates behaviour of duplicate headers.
+
+# Header Text
+
+# Header Text
+
+# header-text
