+++ tests/testsuite/search/reasonable_search_index/src/first/heading-attributes.md
@@ -1,0 +1,5 @@
+# Heading Attributes {#attrs}
+
+## Heading with classes {.class1 .class2}
+
+## Heading with id and classes {#both .class1 .class2}
