+++ tests/testsuite/search/reasonable_search_index/src/first/includes.md
@@ -1,0 +1,3 @@
+# Includes
+
+{{#include ../SUMMARY.md::}}
