+++ tests/testsuite/search/reasonable_search_index/src/first/index.md
@@ -1,0 +1,5 @@
+# First Chapter
+
+more text.
+
+## Some Section
